requirements.txt
Flask==2.3.3
reportlab==4.0.4
python-dotenv==1.1.0
Werkzeug==2.3.7
Pillow==10.0.1
numpy==1.24.3
matplotlib==3.7.2
networkx==3.1
gunicorn==21.2.0
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
setuptools>=65.5.1 # Required for matplotlib 